Added files for BB-labels from label-studio
process_bounding_boxes.py includes functions to handle labeled bounding-boxes that, after being labeled in label-studio, are exported in a .csv file. Includes going from (x,y,w,h)-format to (x,y) for the 4 corners of the BBs (the format used in IC-13/15, relevant for training CRAFT) and visualization. Functions to crop the images in bounding boxes for training the recognition model. Around half of the 1608 observations is arguably noise (i.e irrelevant images in form of missing barcodes (picture of rooftops, bicycles etc.), images with barcodes wholly blurred out/no individual number can be recognized and the barcode itself smeared out, only part of barcode is in the picture and etc.) and have not been labeled.