Implementation of training on CRAFT
Uploaded code for retraining/further training CRAFT - random cropping has been disabled for now as it is suspected that there are some errors associated with it. Repository has been modified from @backtime92's reimplementation from CRAFT (mix of current fork and older) where it is done for SynthData (strongly supervised learning). The function for generating pseudo-labels for weak-supervision and visualizing the psuedo-boundingboxes, we have implemented, together with other changes to the DataLoader. 
Currently validation loss is not calculated at the end of each iteration, and there might not be time to implement this - the way CRAFT calculates loss is a bit technical and coding it while making sure that there are no issues can take time...